# resources/openapi.yml
swagger: "2.0"
info:
  title: BSC API
  description: BSC API
  version: 0.0.1
host: chekov.basecondition.bsc
basePath: /api/v1
...
schemes:
  - https
paths:
...
  /member/timeline:
    get:
      tags:
        - timeline
      summary: Get member timeline sorted descending
      description: Get member timeline sorted descending
      operationId: timeline
      produces:
        - application/json
      responses:
        500:
          $ref: '#/responses/500'
        401:
          $ref: '#/responses/401'
        200:
          description: member timeline
          schema:
            type: array
            items: {}
            description: mixed content (milestones, events, consulting dates ordered by timestamp)
      security:
        - oAuth2: []
...
  /event/{id}:
    get:
      tags:
        - event
      summary: get single event data
      description: get single event data
      operationId: getEvent
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          format: "int64"
      responses:
        500:
          $ref: '#/responses/500'
        403:
          $ref: '#/responses/403'
        401:
          $ref: '#/responses/401'
        400:
          $ref: '#/responses/400'
        200:
          description: returns detailed event information
          schema:
            type: object
            items:
              $ref: '#/definitions/event'
      security:
        - oAuth2: []
...
  /events/subscribe/{id}:
    post:
      tags:
        - event
      summary: subscribe to specified event
      description: subscribe to specified event
      operationId: subscribe
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          format: "int64"
      responses:
        500:
          $ref: '#/responses/500'
        401:
          $ref: '#/responses/401'
        400:
          $ref: '#/responses/400'
        200:
          description: returns detailed event information on success
          schema:
            type: object
            items:
              $ref: '#/definitions/event'
      security:
        - oAuth2: []
  /events/unsubscribe/{id}:
    post:
      tags:
        - event
      summary: unsubscribe from specified event
      description: unsubscribe to specified event
      operationId: unsubscribe
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          format: "int64"
      responses:
        500:
          $ref: '#/responses/500'
        401:
          $ref: '#/responses/401'
        400:
          $ref: '#/responses/400'
        200:
          description: returns detailed event information on success
          schema:
            type: object
            items:
              $ref: '#/definitions/event'
      security:
        - oAuth2: []
...
securityDefinitions:
  basicAuth:
    type: basic
  apiKey:
    type: apiKey
    in: header
    name: X-API-Key
  oAuth2:
    type: oauth2
    flow: accessCode
    authorizationUrl: https://chekov.basecondition.bsc/api/v1/token
    tokenUrl: https://chekov.basecondition.bsc/api/v1/token
    scopes:
      read: Grants read access
      write: Grants write access
      admin: Grants read and write access to administrative information
responses:
  500:
    description: internal_server_error
    schema:
      $ref: '#/definitions/error'
...
  403:
    description: forbidden
    schema:
      $ref: '#/definitions/error'
  401:
    description: invalid_token
    schema:
      $ref: '#/definitions/error'
  400:
    description: invalid_request
    schema:
      $ref: '#/definitions/error'
definitions:
  error:
    type: object
    properties:
      error:
        type: string
      error_description:
        type: string
...
  event:
    type: object
    description: recursive array/object mix (fields specs TODO)
...